# public/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Piyachok API",
    "description": "Документация backend-платформы проекта Piyachok."
  },
  "host": "localhost:3000",
  "basePath": "/api",
...
  "schemes": ["http"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Введіть токен у форматі: Bearer <AccessToken>"
    }
  },
  "paths": {
...
    "/auth/reset-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Скидання пароля (Reset Password)",
        "description": "Приймає одноразовий токен скидання з пошти та новий пароль. Перевіряє валідність токена та його термін дії. У разі успіху оновлює пароль користувача та видаляє токен.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["token", "newPassword"],
              "properties": {
                "token": {
                  "type": "string",
                  "example": "a5e8f49b1c02e3f...",
                  "description": "Секретний токен відновлення, отриманий у листі"
                },
                "newPassword": {
                  "type": "string",
                  "format": "password",
                  "example": "MyNewSecurePassword2026",
                  "description": "Новий пароль користувача (мінімальна довжина на розсуд фронтенду буде захешовано)"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пароль успішно оновлено. Можна виконувати вхід.",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Пароль успішно змінено. Тепер ви можете увійти з новим паролем." }
              }
            }
          },
          "400": {
            "description": "Не передані обов'язкові поля, токен немає або термін дії токена минув",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
...
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Оновлення пари токенів (Refresh Token Rotation)",
        "description": "Приймає старий Refresh Token (з тіла запиту або автоматично з HttpOnly кук), перевіряє його валідність та статус бана користувача. У разі успіху генерує нову пару токенів, перезаписує куку та повертає нові ключі.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "type": "object",
              "properties": {
                "refreshToken": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "description": "Необов'язково. Передається, якщо клієнт не може використовувати cookie HttpOnly автоматично."
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Токени успішно оновлено. Нова кука `refreshToken` встановлена.",
            "headers": {
              "Set-Cookie": {
                "type": "string",
                "description": "Оновлює куку `refreshToken` новими даними."
              }
            },
            "schema": {
              "type": "object",
              "properties": {
                "accessToken": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.new_access_token..."
                },
                "backendRefreshSecretString": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.new_refresh_token...",
                  "description": "Новий Refresh Token для фронтенду"
                }
              }
            }
          },
          "401": {
            "description": "Токен відсутній, недійсний, прострочений, або користувач забанен/удален",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
...
    "/messages": {
      "post": {
        "tags": ["Messages"],
        "summary": "Надсилання повідомлення до закладу (публічно)",
        "description": "Дозволяє будь-якому користувачеві або гостю надіслати повідомлення конкретному закладу (наприклад, питання про наявність місць, бронювання чи пропозицію).",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["venueId", "name", "email", "content"],
              "properties": {
                "venueId": { "type": "string", "example": "venue-456-uuid", "description": "ID заведения" },
                "name": { "type": "string", "example": "Олена", "description": "Имя отправителя" },
                "email": { "type": "string", "format": "email", "example": "olena@example.com", "description": "Контактная почта" },
                "content": { "type": "string", "example": "Чи можна замовити столик на терасі на цю п'ятницю?", "description": "Текст сообщения" },
                "type": { "type": "string", "enum": ["QUESTION", "BOOKING", "FEEDBACK"], "default": "QUESTION", "example": "QUESTION", "description": "Тип сообщения" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Повідомлення успішно створено та відправлено менеджеру",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Повідомлення успішно відправлено! Менеджер зв’яжется з вами." },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "example": "msg-123-uuid" },
                    "venueId": { "type": "string", "example": "venue-456-uuid" },
                    "name": { "type": "string", "example": "Олена" },
                    "email": { "type": "string", "example": "olena@example.com" },
                    "content": { "type": "string", "example": "Чи можна замовити столик на терасі..." },
                    "type": { "type": "string", "example": "QUESTION" },
                    "createdAt": { "type": "string", "format": "date-time", "example": "2026-07-15T12:00:00.000Z" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Пропущено обов'язкові поля",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "404": {
            "description": "Заклад із зазначеним ID не знайдено",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      },
      "get": {
        "tags": ["Messages"],
        "summary": "Отримання списку повідомлень (Для менеджерів та адмінів)",
        "description": "Повертає вхідні повідомлення. Звичайні користувачі отримують 403. Менеджери бачать повідомлення лише для своїх закладів. Адміністратори бачать усі повідомлення системи.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "Фільтр типу повідомлення (наприклад, BOOKING, QUESTION)"
          }
        ],
        "responses": {
          "200": {
            "description": "Список повідомлень успішно отримано",
            "schema": {
              "type": "object",
              "properties": {
                "messages": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "example": "msg-123-uuid" },
                      "venueId": { "type": "string", "example": "venue-456-uuid" },
                      "name": { "type": "string", "example": "Олена" },
                      "email": { "type": "string", "example": "olena@example.com" },
                      "content": { "type": "string", "example": "Чи можна замовити столик..." },
                      "type": { "type": "string", "example": "QUESTION" },
                      "createdAt": { "type": "string", "format": "date-time" },
                      "venue": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string", "example": "П'яний Лис" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Неавторизований доступ (потрібний токен)",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "403": {
            "description": "Доступ заборонено (роль користувача USER)",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
...
    "/tags": {
      "get": {
        "tags": ["Tags"],
        "summary": "Отримання списку всіх тегів",
        "description": "Повертає повний список тегів, відсортованих за алфавітом (A-Z, А-Я). Доступно публічно.",
        "responses": {
          "200": {
            "description": "Список тегів успішно отримано",
            "schema": {
              "type": "object",
              "properties": {
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "example": "tag-111-uuid" },
                      "name": { "type": "string", "example": "Крафт" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      },
      "post": {
        "tags": ["Tags"],
        "summary": "Створення нового тегу",
        "description": "Дозволяє авторизованому користувачу створити новий тег. Назва тегу має бути унікальною.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name"],
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Жива музика",
                  "description": "Назва тегу (буде автоматично очищена від зайвих пробілів)"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Тег успішно створено",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Тег успішно створено" },
                "tag": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "example": "tag-222-uuid" },
                    "name": { "type": "string", "example": "Жива музика" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Не передано назву тегу або такий тег вже існує в системі",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "401": {
            "description": "Неавторизований доступ",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
...
    "/user/me": {
      "get": {
        "tags": ["Users"],
        "summary": "Отримання профілю поточного користувача",
        "description": "Повертає інформацію про авторизованого користувача (ID, ім'я, email, роль, аватар, телефон) на основі переданого Bearer токена.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Дані профілю успішно отримані",
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "example": "user-789-uuid" },
                    "name": { "type": "string", "example": "Дмитро" },
                    "email": { "type": "string", "format": "email", "example": "dmitro@example.com" },
                    "role": { "type": "string", "example": "MANAGER", "description": "Роль користувача (USER, MANAGER, ADMIN, SUPERADMIN)" },
                    "avatarUrl": { "type": "string", "x-nullable": true, "example": "https://my-bucket.s3.amazonaws.com/avatars/my-photo.jpg" },
                    "phoneNumber": { "type": "string", "x-nullable": true, "example": "+380671234567" },
                    "createdAt": { "type": "string", "format": "date-time", "example": "2026-01-15T10:30:00.000Z" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Неавторизований доступ (невірний або прострочений токен)",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "404": {
            "description": "Користувача з ID з токена не знайдено в базі даних",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "message": { "type": "string", "example": "Внутрішня помилка сервера / Доступ заборонено" }
      }
    },
...
  }
}